Change type

+++ src/BmiService.java
@@ -1,5 +1,5 @@
 public class BmiService {
-    public double calculate (double height, double mass) {
+    public int calculate (double height, double mass) {
         int bodyMassIndex = (int) (mass / (height * height));
         return bodyMassIndex;
     }
